requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.6
elementpath==3.0.2
et-xmlfile==1.1.0
idna==3.3
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
regex==2021.4.4
requests==2.26.0
six==1.16.0
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.7
xmlschema==2.1.1
